+++ config/settings/ci.py
@@ -1,5 +1,8 @@
+from tempfile import TemporaryDirectory
+
 from .base import *  # NOQA
 
+MEDIA_ROOT = TemporaryDirectory().name
 SECRET_KEY = 'SecretKeyForUseOnTravis'
 DEBUG = False
 TEMPLATE_DEBUG = True